jdk1.8.0_40 and Other Things
This update had to be done, especially for me. I updated from 1.8.0_31
or something like that to 40, which was recently released (today maybe).
Main change is in Main.java, where I had to change the value of
MY_JAVA_LOC. Also, I figured out where the Forms thing came from and
removed it (as we don't need it here as it turns out). There is a major
issue when running the program, but it almost works. Will open issue.